Browser support fixes and invalidValue validation
- When Chrome finds an error on a form, the submit event is not fired

Listening to `click` on `[type=submit]` (annoyingly enough) seems to
be better than listening to submit events. Why do we have a click
event instead of an "attempt-to-submit" event? I blame onClick
handlers and bad coffee.

- New validation test: Invalid Value.

This lets us add additional validation tests to a field where typical validation passes. For
example if you have a username, you may require it to be a certain
length, if that is satisfied, you may also check to see if it is in
use. If a username is in use, you can set `dataset.invalidValue = input.value`
Then set `dataset.invalidValueMessage="This username is already in use"
This will allow current validation to continue to work, while also
notifying the user that this specific value cannot be used.